Alteração do tipo do atributo de amount

+++ service-provision/src/main/java/com/api/sale/dtos/LaborDto.java
@@ -4,6 +4,9 @@ import java.io.Serializable;
 import java.math.BigDecimal;
 import java.time.LocalDate;
 
+import com.api.sale.entities.Client;
+import com.api.sale.entities.Labor;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -15,7 +18,7 @@ public class LaborDto implements Serializable{
 
     private String description;
     private Integer idClient;
-    private BigDecimal amount;
+    private Double amount;
     private LocalDate date;
     
 }
